Functions to retrieve variables for the TempSED model.

TempSED_get0Dvars, TempSED_get1Dvars, TempSED_getstates retrieve the (0-dimensional or 1-dimensional) output variables or the state variables of TempSED model solutions. When called with out not specified, will return the names of these variables and their units.

TempSED_get0Dvars(out, as.vector = FALSE, which = NULL) 
  TempSED_get1Dvars(out, which = NULL) 
  TempSED_getstates(out, which = NULL)

Arguments

out

an output object returned by TempSED_run1D.

as.vector

if TRUE will return the parameter vector, else a data.frame that also contains the units.

which

if not NULL, a vector with names of the variables/parameters to return.

Value

TempSED_get0Dvars and TempSED_get1Dvars return the output variables of the solution as a vector or data.frame. For dynamic runs, the output is averaged over the mean of the run.

TempSED_get1Dvars always returns the sediment depth and the porosity as the first two columns.

Author

Karline Soetaert

References

Soetaert K, PMJ Herman and JJ Middelburg, 1996a. A model of early diagenetic processes from the shelf to abyssal depths. Geochimica Cosmochimica Acta, 60(6):1019-1040.

Soetaert K, PMJ Herman and JJ Middelburg, 1996b. Dynamic response of deep-sea sediments to seasonal variation: a model. Limnol. Oceanogr. 41(8): 1651-1668.

See also

Examples


# defaults
  TempSED_getstates()
#>         names units                  description
#> 1 Temperature   dgC Temperature of bulk sediment
  head(TempSED_get0Dvars())
#>              names values units                    description
#> 1 Mean_temperature     NA   dgC      Mean sediment temperature
#> 2 Watertemperature     NA   dgC Temperature of overlying water
#> 3      Waterheight     NA     m      Height of overlying water
#> 4   Airtemperature     NA   dgC             Temperature of air
#> 5      Airhumidity     NA     -       Relative humidity of air
#> 6         Pressure     NA    Pa                   Air pressure
  TempSED_get1Dvars()
#>         names units                  description
#> 1 Temperature   dgC Temperature of bulk sediment
  
# some runs to work with  
  defdyn    <- TempSED_run1D()

# altered run
  out <- TempSED_run1D(parms = list(em_air = 0.95))

# 0-D outputs
  cbind(dyna   = TempSED_get0Dvars(defdyn, as.vector = TRUE), 
        ref    = TempSED_get0Dvars(out))
#>                              dyna           ref.names    ref.values ref.units
#> Mean_temperature     9.999140e+00    Mean_temperature  9.999140e+00       dgC
#> Watertemperature     1.000000e+01    Watertemperature  1.000000e+01       dgC
#> Waterheight          0.000000e+00         Waterheight  0.000000e+00         m
#> Airtemperature       1.000000e+01      Airtemperature  1.000000e+01       dgC
#> Airhumidity          3.000000e-01         Airhumidity  3.000000e-01         -
#> Pressure             1.013250e+05            Pressure  1.013250e+05        Pa
#> Solarradiation       1.000000e+02      Solarradiation  1.000000e+02      W/m2
#> Windspeed            1.000000e+00           Windspeed  1.000000e+00       m/s
#> RadiationSWI         8.500000e+01        RadiationSWI  8.500000e+01      W/m2
#> Evaporationrate      8.877979e-06     Evaporationrate  8.877979e-06   kg/m2/s
#> Heatflux_air        -1.000661e+02        Heatflux_air -1.000661e+02      W/m2
#> Heatflux_total      -1.506608e+01      Heatflux_total -1.506608e+01      W/m2
#> Heatflux_convection  0.000000e+00 Heatflux_convection  0.000000e+00      W/m2
#> Heatflux_radiation   8.500000e+01  Heatflux_radiation  8.500000e+01      W/m2
#> Heatflux_latent     -2.200206e+01     Heatflux_latent -2.200206e+01      W/m2
#> Heatflux_sensible    9.976988e-01   Heatflux_sensible  9.976988e-01      W/m2
#> Heatflux_backrad    -7.906172e+01    Heatflux_backrad -7.906172e+01      W/m2
#> Total_irrigation     0.000000e+00    Total_irrigation  0.000000e+00  m3.dgC/s
#>                                                 ref.description
#> Mean_temperature                      Mean sediment temperature
#> Watertemperature                 Temperature of overlying water
#> Waterheight                           Height of overlying water
#> Airtemperature                               Temperature of air
#> Airhumidity                            Relative humidity of air
#> Pressure                                           Air pressure
#> Solarradiation                      Solar radiation (shortwave)
#> Windspeed                                            Wind speed
#> RadiationSWI                Solar radiation at sediment surface
#> Evaporationrate                             Rate of evaporation
#> Heatflux_air              Total heat flux at interface with air
#> Heatflux_total      Total heat added (upper boundary+radiation)
#> Heatflux_convection  Convective heat flux with overlying water 
#> Heatflux_radiation               Shortwave radiation heat input
#> Heatflux_latent                latent heat input (~evaporation)
#> Heatflux_sensible        sensible heat input (~air temperature)
#> Heatflux_backrad      net longwave radiation (net backradation)
#> Total_irrigation            temperature input due to irrigation

# 1-D outputs
  head(TempSED_get1Dvars(out))
#>             x por Temperature
#> 1 0.000498923 0.5    9.601291
#> 2 0.001530360 0.5    9.643470
#> 3 0.002631242 0.5    9.667866
#> 4 0.003806243 0.5    9.685699
#> 5 0.005060354 0.5    9.701432
#> 6 0.006398901 0.5    9.716696
  head(TempSED_get1Dvars(defdyn))
#>             x por Temperature
#> 1 0.000498923 0.5    9.601291
#> 2 0.001530360 0.5    9.643470
#> 3 0.002631242 0.5    9.667866
#> 4 0.003806243 0.5    9.685699
#> 5 0.005060354 0.5    9.701432
#> 6 0.006398901 0.5    9.716696