Properties of the major sediment-constituent minerals.

air_properties calculates the humidity, density and vapor pressure of the air.

mineral_properties (
    mineral = c("Quartz", "Orthoclase", "Albite", "Anorthite",  "Calcite" ,                  
                "Muscovite", "Illite", "Smectite (montmorillonite)", 
                "Chlorite", "Pyrite"))
mineral_density    (
    mineral = c("Quartz", "Orthoclase", "Albite", "Anorthite",  "Calcite" ,                  
                "Muscovite", "Illite", "Smectite (montmorillonite)", 
                "Chlorite", "Pyrite"))

mineral_tc   (
    mineral = c("Quartz", "Orthoclase", "Albite", "Anorthite",  "Calcite" ,                  
                "Muscovite", "Illite", "Smectite (montmorillonite)", 
                "Chlorite", "Pyrite"))

mineral_cp         (    
    mineral = c("Quartz", "Orthoclase", "Albite", "Anorthite",  "Calcite" ,                  
                "Muscovite", "Illite", "Smectite (montmorillonite)", 
                "Chlorite", "Pyrite"))

mineral_td         (
    mineral = c("Quartz", "Orthoclase", "Albite", "Anorthite",  "Calcite" ,                  
                "Muscovite", "Illite", "Smectite (montmorillonite)", 
                "Chlorite", "Pyrite"))

Arguments

mineral

Sediment-constituent mineral, one element or a vector.

Value

function mineral_properties returns a list with:

density, the mineral density, [kg/m3]

cp, the specific heat capacity of the mineral(s), [J/kg/dg]

td, the thermal conductivity of the mineral(s), [W/m/dg]

tc, the thermal diffusivity of the mineral(s), [m2/s]

The object attributes contain:

A description of these variables, including their units

use attibutes(object)$description to extract it.

Author

Karline Soetaert

References

Shusaku Goto, Makoto Yamano, Sumito Morita, Toshiya Kanamatsu, Akihiro Hachikubo, Satsuki Kataoka, Manabu Tanahashi, Ryo Matsumoto, 2017. Physical and thermal properties of mud-dominant sediment from the Joetsu Basin in the eastern margin of the Japan Sea. Mar Geophys Res 38:393-407, DOI 10.1007/s11001-017-9302-y

See also

Examples


mineral_properties()
#>                       mineral density    tc       td  cp
#> 1                      Quartz    2648  7.69 3.92e-06 741
#> 2                  Orthoclase    2570  2.32 1.28e-06 707
#> 3                      Albite    2620  2.14 1.05e-06 776
#> 4                   Anorthite    2760  1.68 8.17e-07 745
#> 5                     Calcite    2710  3.59 1.62e-06 820
#> 6                   Muscovite    2831  2.32 1.03e-06 796
#> 7                      Illite    2660  1.85 8.61e-07 808
#> 8  Smectite (montmorillonite)    2608  1.88 9.07e-07 795
#> 9                    Chlorite    2800  5.15 2.25e-06 818
#> 10                     Pyrite    5011 19.21 7.40e-06 518
mineral_properties("Quartz")
#>   mineral density   tc       td  cp
#> 1  Quartz    2648 7.69 3.92e-06 741

mineral_tc()
#>                       mineral    tc
#> 1                      Quartz  7.69
#> 2                  Orthoclase  2.32
#> 3                      Albite  2.14
#> 4                   Anorthite  1.68
#> 5                     Calcite  3.59
#> 6                   Muscovite  2.32
#> 7                      Illite  1.85
#> 8  Smectite (montmorillonite)  1.88
#> 9                    Chlorite  5.15
#> 10                     Pyrite 19.21
mineral_tc("Quartz")
#>   mineral   tc
#> 1  Quartz 7.69

# description
attributes(mineral_properties())$description
#>     names          description   units
#> 1 density              density   kg/m3
#> 2      tc thermal conductivity   W/m/s
#> 3      td  thermal diffusivity    m2/s
#> 4      cp        heat capacity J/kg/dg